# [bash-completion] Optimize startup + two minor fixes