@junegunn 2020-11-12
Thanks Tomáš, I really like what you did here. Bravo.