@wingpin6 2020-11-13
Hi guys, thanks for your efforts on this program and improving it.

Unfortunately for me, a commit within this PR has caused my terminal to crash when trying to tab. I've identified commit ef2c29d5d4974cc773149d368488a6d8a0fcc1fc through git bisect as the offending commit.

For reference I am currently using:
GNU bash, version 5.0.17(1)-release (x86_64-pc-linux-gnu)
on Ubuntu 20.04
Terminal choice does not seem to matter; I crash on both Alacritty (my daily driver) and Gnome Terminal.

Where would I find logs on how to diagnose this issue?