@liskin 2020-11-13
@wingpin6 Could you please try running a nested `bash` in your terminal and reproducing this in it? This way the terminal won't terminate when the bash terminates and we might see an error message if there is any.