@wingpin6 2020-11-13
@liskin Thank you for the tip; after doing nested bash --norc I found that
- sourcing only .bashrc did not crash
- sourcing only .fzf.bash did not crash
- sourcing .fzf.bash and then .bashrc did not crash
- However, sourcing .bashrc and then .fzf.bash caused it to crash

The error it outputted was Segmentation fault.

I looked deeper into this issue, and was able to fix it if I removed
`[ -f ~/.fzf.bash ] && source ~/.fzf.bash`
from my .bashrc

It seems something in that commit mentioned earlier may have sourced my .bashrc, which sources .fzf.bash and then causes an infinite loop?

I was able to fix my issue by not using .bashrc as my default bash config file, and then sourcing .bashrc and .fzf.bash within that new config file separately. 

Let me know if you need more info or if you can't replicate that.