@liskin 2020-11-13
@wingpin6 Segmentation fault seems to be a bug in bash, provided it comes from bash and not from something else.

Could you get a stack trace from the core dump? I'm not sure if Ubuntu has systemd-coredump, but if yes, then this should be as easy as running `coredumpctl info` and sending the output. If you could also `apt install bash-dbgsym` (https://wiki.ubuntu.com/Debug%20Symbol%20Packages) and then do `coredumpctl debug` and then type `bt`into gdb, that would be wonderful.

If there's no systemd-coredump in Ubuntu, instead of running bash directly, run `gdb -ex run --args bash` and reproduce the crash and then when it crashes, get stack trace by typing `bt`.

I also looked at recent changes in bash in Ubuntu and it seems that this may possibly help: https://github.com/liskin/fzf/commit/6b4ab0a2af42c2103cb4b7b96f4fba076b4cd002, although that is really a wild guess, nothing else.