@liskin 2020-11-13
Oh, and it seems I can replicate that now. The key is sourcing completion.bash twice!
No need to debug it then, I'll take care of it!