@liskin 2020-11-13
@wingpin6 Fix in https://github.com/junegunn/fzf/pull/2250. Thanks for reporting this and sorry for breaking your setup.

(You might, however, consider sourcing completion.bash just once in the future, as that will be faster. :-))