@wingpin6 2020-11-13
Thank you for the quick fix!

I agree this case is most likely uncommon. I have a personal config file that I use on all my machines and a work config file for my work machine. I source both of them with work config overriding anything I defined in my personal config, with some overlaps.

I happened to source .fzf.bash in both of them; not sure how that got into my work config file. But great to find out that I can remove that line now!